+++ king_phisher/client/dialogs/configuration.py
@@ -138,7 +138,6 @@ class ConfigurationDialog(gui_utilities.GladeGObject):
 			'entry_ssh_server',
 			'entry_ssh_username',
 			# Client Tab
-			'checkbutton_remove_attachment_metadata',
 			'combobox_spf_check_level',
 			# Plugins Tab
 			'box_plugin_options'