+++ king_phisher/client/tabs/mail.py
@@ -37,10 +37,8 @@ import os
 import re
 import sys
 import urllib
-import zipfile
 
 from king_phisher import its
-from king_phisher import scrubber
 from king_phisher import spf
 from king_phisher import utilities
 from king_phisher.client import dialogs
@@ -138,16 +136,7 @@ class MailSenderSendTab(gui_utilities.GladeGObject):
 			gui_utilities.show_dialog_warning('Invalid Attachment', self.parent, 'The specified attachment file can not be read.')
 			return False
 		self.text_insert("File '{0}' will be attached to sent messages.\n".format(os.path.basename(attachment)))
-		_, extension = os.path.splitext(attachment)
-		extension = extension[1:]
-		if self.config['remove_attachment_metadata'] and extension in ('docm', 'docx', 'pptm', 'pptx', 'xlsm', 'xlsx'):
-			self.text_insert('Attachment file detected as MS Office 2007+, ')
-			try:
-				scrubber.remove_office_metadata(attachment)
-			except zipfile.BadZipfile:
-				self.text_insert('failed to remove metadata.\n')
-			else:
-				self.text_insert('metadata has been removed.\n')
+
 		md5 = hashlib.new('md5')
 		sha1 = hashlib.new('sha1')
 		sha256 = hashlib.new('sha256')